requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.ini
#
ansible==2.7.1
appnope==0.1.0            # via ipython
asn1crypto==0.24.0        # via cryptography
backcall==0.1.0           # via ipython
bcrypt==3.1.4             # via paramiko
cffi==1.11.5              # via bcrypt, cryptography, pynacl
click==7.0                # via pip-tools
cryptography==2.4.1       # via ansible, paramiko
decorator==4.3.0          # via ipython, traitlets
django-debug-toolbar==1.10.1
django==2.1.3
flake8==3.6.0
idna==2.7                 # via cryptography
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1            # via ipdb
isort==4.3.4
jedi==0.13.1              # via ipython
jinja2==2.10              # via ansible
markupsafe==1.1.0         # via jinja2
mccabe==0.6.1             # via flake8
paramiko==2.4.2           # via ansible
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.3.0
pip-tools==3.1.0
prompt-toolkit==2.0.7     # via ipython
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0         # via pexpect
pyasn1==0.4.4             # via paramiko
pycodestyle==2.4.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.0.0           # via flake8
pygments==2.2.0           # via ipython
pynacl==1.3.0             # via paramiko
pytz==2018.7              # via django
pyyaml==3.13              # via ansible
six==1.11.0               # via bcrypt, cryptography, pip-tools, prompt-toolkit, pynacl, traitlets
sqlparse==0.2.4           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
